refactor: 도메인 계층 모델과 데이터 계층 모델 간 변환 담당 함수들을 데이터 계층으로 이전
