+++ app/src/main/java/woowacourse/movie/data/mapper/TheaterMapper.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain.mapper
+package woowacourse.movie.data.mapper
 
 import woowacourse.movie.data.database.theater.TheaterEntity
 import woowacourse.movie.domain.Theater
@@ -9,11 +9,4 @@ fun TheaterEntity.toTheater() =
         id = id,
         name = name,
         screeningTimes = screeningTimes.map { LocalTime.parse(it) },
-    )
-
-fun Theater.toTheaterEntity() =
-    TheaterEntity(
-        id = id,
-        name = name,
-        screeningTimes = screeningTimes.map(LocalTime::toString),
     )
